Start of the notes repository.
This is just to keep all the content at the same
place.

I'm still looking at Jekyll to present this in a nice
fashion, however content is content, and I like to
have it separated from the rest.
Note, I might reupload this on others websites, translated
or not (Medium, Patreon, Note, Qiita, ...).

Signed-off-by: Voyage <voyage@miouyouyou.fr>